Split file format spec from import/export functionality
