+++ file-format.ts
@@ -1,63 +1,12 @@
 import {Profile, CallTreeNode, Frame, CallTreeProfileBuilder, FrameInfo} from './profile'
 import {TimeFormatter, ByteFormatter, RawValueFormatter} from './value-formatters'
 import {lastOf} from './utils'
-
-interface SerializedFrame {
-  name: string
-  file?: string
-  line?: number
-  col?: number
-}
-
-interface SerializedNode {
-  // Index into the frames array on the SerializedProfile
-  frame: number
-
-  // Index into the nodes array on the SerializedProfile
-  parent?: number
-}
-
-export type WeightUnit =
-  | 'none'
-  | 'nanoseconds'
-  | 'microseconds'
-  | 'milliseconds'
-  | 'seconds'
-  | 'bytes'
-
-interface SerializedSamplingProfile {
-  // Type of profile. This will future proof the file format to allow many
-  // different kinds of profiles to be contained and each type to be part of
-  // a discriminate union.
-  type: 'SamplingProfile'
-
-  // Name of the profile. Typically a filename for the source of the profile.
-  name: string
-
-  // List of all call frames
-  frames: SerializedFrame[]
-
-  // List of nodes in the call tree
-  nodes: SerializedNode[]
-
-  // List of indices into nodes, with -1 indicating that the call-stack
-  // was empty at the time of the sample
-  samples: number[]
-
-  // The weight of the sample at the given index. Should have
-  // the same length as the samples array.
-  weights: number[]
-
-  // Unit of the weights provided in the profile. If none provided,
-  // the weights are assumed to be unit-less.
-  weightUnit: WeightUnit
-}
-
-export interface SerializedSpeedscopeFile {
-  version: string
-  exporter: 'https://www.speedscope.app'
-  profiles: SerializedSamplingProfile[]
-}
+import {
+  SerializedSpeedscopeFile,
+  SerializedSamplingProfile,
+  SerializedFrame,
+  SerializedNode,
+} from './file-format-spec'
 
 export interface InMemorySpeedscopeFile {
   version: string
